Combine and Remove unused Base Classes
